+++ src/Components/Business/RelatedBusinessList.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import RelatedBusiness from "./RelatedBusiness";
+
+export default () => (
+  <>
+    <h3 className="text-uppercase text-muted h6 my-3">Related Businesses</h3>
+    <RelatedBusiness />
+    <RelatedBusiness />
+    <RelatedBusiness />
+    <RelatedBusiness />
+    <RelatedBusiness />
+  </>
+);
